hub_server/requirements.txt
argparse
uvicorn[standard]
gunicorn
fastapi
opencv_python
redis
celery[redis]
flower
PyYAML
timeago
requests
urllib3==1.26.6
pymongo
httpx
prometheus-client
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-logging
opentelemetry-exporter-otlp
opentelemetry-api
opentelemetry-sdk
celery
clique
imutils
